Rename application to service Rename infra to repository
