+++ dayco-uaa/src/main/java/io/github/dayco/uaa/auth/config/UaaWebSecurityConfiguration.java
@@ -21,8 +21,8 @@ import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import io.github.dayco.uaa.auth.application.UaaUserDetailService;
-import io.github.dayco.uaa.manager.application.ManagerService;
+import io.github.dayco.uaa.auth.service.UaaUserDetailService;
+import io.github.dayco.uaa.manager.service.ManagerService;
 
 @Configuration
 @EnableWebSecurity
