+++ dayco-uaa/src/main/java/io/github/dayco/uaa/user/repository/UserJpaRepository.java
@@ -1,4 +1,4 @@
-package io.github.dayco.uaa.user.infra;
+package io.github.dayco.uaa.user.repository;
 
 import java.util.Optional;
