--- dayco-uaa/src/test/java/io/github/dayco/uaa/FileUploader.java
@@ -1,4 +1,0 @@
-package io.github.dayco.uaa;
-
-public class FileUploader {
-}
